fix: auth, session, OneDrive and login routing bugs
- fix(auth): await failed-login attempt increment to prevent race on fast retries bypassing lockout
- fix(auth): dev cookie maxAge was in seconds not ms (43 min instead of 30 days)
- fix(auth/bff): rewrite Secure;SameSite=None cookies to SameSite=Lax in local dev so session works on HTTP
- fix(login): server-side redirect on /login always sent admins to /mail; now decodes JWT to route correctly
- fix(login): checkAuth useEffect checked non-existent `authenticated` field; now reads whoami response correctly
- fix(login): add client-side email format validation before fetch; safe JSON parse to handle HTML error responses
- fix(useAuthedSWR): 401 logged as console.error triggering dev overlay; downgrade to warn and redirect to /login
- fix(onedrive): clear cached Graph token on 401/403 so next retry fetches a fresh token
- fix(db-schema): locked_until INTEGER -> BIGINT to match production postgres schema
- feat(frontend): add unlock-account BFF route and DevApiCheckClient split

+++ apps/backend/scripts/db-schema.sql
@@ -20,7 +20,7 @@ CREATE TABLE IF NOT EXISTS user (
   email_verified_at INTEGER,
   status        TEXT DEFAULT 'active',
   login_attempts INTEGER DEFAULT 0,
-  locked_until  INTEGER,
+  locked_until  BIGINT,
   created_at    INTEGER NOT NULL,
   updated_at    INTEGER NOT NULL
 );
